Moves transactions table summary logic from generic db_query to transactions_manager and removes db_query + test module.
